Added center heart image
Added the image I made of the heart starting with one line and ending with three lines to symbolize one donation saving three lives.
The image scales to show the complete image. It goes from one end of the screen to another so that the lines look like they're continuing.

Black image uploaded in case that version needs to be tried during testing.