Changed home screen text
Updated the home screen text.

Changed the order of the text (slogan, heart image, facts).

Changed the size of h1 and p to be more readable on screen and make better use of screen space.

Changed p text to bold to match the style of the rest of the page.