Created the story page
Created the page where the video stories will be displayed. The page exists of a banner with a featured video and a list of available videos to watch. The videos have a header, a subheader and an image.
Added the story page to the router and menu.
Changed the names of the home page buttons.
Updated css file with better structure and comments explaining the specific part of the css.